# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kogda-Virastu API ",
    "description": "API бэкенд-сервиса проекта \"КогдаВырасту\"",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.kogda-virastu.com/v1"
    },
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "paths": {
...
    "/articles/": {
      "get": {
        "tags": [
          "article"
        ],
        "summary": "Получить список статей",
        "description": "",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "favorited",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "articles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ArticleData"
                      }
                    },
                    "articlesCount": {
                      "type": "number",
                      "example": 1
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "articles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ArticleData"
                      }
                    },
                    "articlesCount": {
                      "type": "number",
                      "example": 1
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "article"
        ],
        "summary": "Cоздать новую публикацию",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "$ref": "#/components/schemas/Article"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "$ref": "#/components/schemas/Article"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "article": {
                    "$ref": "#/components/schemas/ArticleData"
                  }
                }
              }
            },
            "application/xml": {
              "schema": {
                "type": "object",
                "properties": {
                  "article": {
                    "$ref": "#/components/schemas/ArticleData"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/articles/feed": {
      "get": {
        "tags": [
          "article"
        ],
        "summary": "Получить фид публикаций",
        "description": "",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "articles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ArticleData"
                      }
                    },
                    "articlesCount": {
                      "type": "number",
                      "example": 1
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "articles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ArticleData"
                      }
                    },
                    "articlesCount": {
                      "type": "number",
                      "example": 1
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/articles/{article}": {
      "get": {
        "tags": [
          "article"
        ],
        "summary": "Получить публикацию с указанным Id",
        "description": "",
        "parameters": [
          {
            "name": "article",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "$ref": "#/components/schemas/Article"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "$ref": "#/components/schemas/Article"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "article"
        ],
        "summary": "Обновить публикацию",
        "description": "",
        "parameters": [
          {
            "name": "article",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "$ref": "#/components/schemas/Article"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "$ref": "#/components/schemas/Article"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "article": {
                    "$ref": "#/components/schemas/ArticleData"
                  }
                }
              }
            },
            "application/xml": {
              "schema": {
                "type": "object",
                "properties": {
                  "article": {
                    "$ref": "#/components/schemas/ArticleData"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "article"
        ],
        "summary": "Удалить публикацию с указанным Id",
        "description": "",
        "parameters": [
          {
            "name": "article",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/articles/{article}/favorite": {
      "post": {
        "tags": [
          "article"
        ],
        "summary": "Добавить публикацию в избранное",
        "description": "",
        "parameters": [
          {
            "name": "article",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "$ref": "#/components/schemas/Article"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "$ref": "#/components/schemas/Article"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "article"
        ],
        "summary": "Удалить публикацию из избранного",
        "description": "",
        "parameters": [
          {
            "name": "article",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "$ref": "#/components/schemas/Article"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "article": {
                      "$ref": "#/components/schemas/Article"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/articles/{article}/comments": {
      "get": {
        "tags": [
          "article"
        ],
        "summary": "Получить комментарии статьи",
        "description": "",
        "parameters": [
          {
            "name": "article",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Comment"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Comment"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "article"
        ],
        "summary": "Добавить новый комментарий к статье",
        "description": "",
        "parameters": [
          {
            "name": "article",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            },
            "application/xml": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/articles/{article}/comments/{comment}": {
      "delete": {
        "tags": [
          "article"
        ],
        "description": "",
        "parameters": [
          {
            "name": "article",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "slug публикации"
          },
          {
            "name": "comment",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id комментария"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "amazinguser"
          },
          "nickname": {
            "type": "string",
            "example": "tiredcoder"
          },
          "bio": {
            "type": "string",
            "example": "Something about me"
          },
          "image": {
            "type": "string",
            "example": "http://ya.ru/avatar/1.jpg"
          },
          "following": {
            "type": "boolean",
            "example": false
          },
          "followingTags": {
            "type": "array",
            "example": [],
            "items": {}
          }
        },
        "xml": {
          "name": "UserProfile"
        }
      },
...
      "ArticleData": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Тестовая запись"
          },
          "description": {
            "type": "string",
            "example": "О чем-то интересном"
          },
          "body": {
            "type": "string",
            "example": "А это текст статьи 2"
          },
          "tagList": {
            "type": "array",
            "example": [
              "теги",
              "никто",
              "не",
              "читает"
            ],
            "items": {
              "type": "string"
            }
          },
          "link": {
            "type": "string",
            "example": "https://static.tildacdn.com/tild3435-3464-4534-b032-633331373365/Main_3.jpg"
          }
        },
        "xml": {
          "name": "ArticleData"
        }
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "62a3392219fbd7ec26ff90a4"
          },
          "body": {
            "type": "string",
            "example": "Тест сообщения"
          },
          "createdAt": {
            "type": "string",
            "example": "2022-06-10T12:29:22.830Z"
          },
          "state": {
            "xml": {
              "name": "publishstate"
            },
            "$ref": "#/components/schemas/PublishState"
          },
          "author": {
            "xml": {
              "name": "userprofile"
            },
            "$ref": "#/components/schemas/UserProfile"
          }
        },
        "xml": {
          "name": "Comment"
        }
      },
      "Article": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "testovaya-zapis-szpp6o"
          },
          "title": {
            "type": "string",
            "example": "Тестовая запись"
          },
          "description": {
            "type": "string",
            "example": "О чем-то интересном"
          },
          "body": {
            "type": "string",
            "example": "А это текст статьи 2"
          },
          "state": {
            "xml": {
              "name": "publishstate"
            },
            "$ref": "#/components/schemas/PublishState"
          },
          "createdAt": {
            "type": "string",
            "example": "2022-06-07T00:20:09.388Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2022-06-13T12:49:06.305Z"
          },
          "tagList": {
            "type": "array",
            "example": [
              "теги",
              "никто",
              "не",
              "читает"
            ],
            "items": {
              "type": "string"
            }
          },
          "favorited": {
            "type": "boolean",
            "example": false
          },
          "favoritesCount": {
            "type": "number",
            "example": 5
          },
          "author": {
            "xml": {
              "name": "userprofile"
            },
            "$ref": "#/components/schemas/UserProfile"
          }
        },
        "xml": {
          "name": "Article"
        }
      },
...
      "PublishState": {
        "type": "string",
        "enum": [
          "published",
          "pending",
          "declined"
        ],
        "xml": {
          "name": "PublishState"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {"bearerAuth": []}
  ]
}